ace_backend/requirements.txt
asn1crypto==0.24.0
beautifulsoup4==4.7.1
certifi==2018.11.29
cffi==1.12.2
chardet==3.0.4
cryptography==2.6.1
Django==2.1.7
django-cors-headers==2.4.0
django-filter==2.1.0
django-rest-knox==3.6.0
djangorestframework==3.9.1
djangorestframework-stubs==0.4.0
drf-nested-routers==0.91
entrypoints==0.3
flake8==3.7.4
idna==2.8
Markdown==3.0.1
mccabe==0.6.1
parse==1.11.1
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.0
Pygments==2.3.1
pyOpenSSL==19.0.0
pytz==2018.9
requests==2.21.0
rope==0.11.0
six==1.12.0
soupsieve==1.7.3
urllib3==1.24.1
